requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==24.2.0
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
pytz==2024.2
six==1.16.0
typing_extensions==4.12.2
tzlocal==5.2
yarl==1.16.0
